Moved performance to evaluation
